+++ joystick.c
@@ -35,7 +35,7 @@
 #include "machine.h"
 #include "joystick.h"
 
-#include "plugins/twilighte_card/oric_twilighte_board_plugin.h"
+#include "plugins/twilighte_board/oric_twilighte_board_plugin.h"
 
 struct keyjoydef
 {
@@ -301,13 +301,13 @@ void joy_buildmask( struct machine *oric )
 
     if( gimme_port_a )
     {
-      //twilighteboard_oric_write(oric->twilighte, 0x321, 0xff,mkmask_f);
+      //twilighteboard_oric_write(oric->twilighte, 0x321, 0xff, mkmask_f);
     }
 
     if( gimme_port_b )
     {
         // port B
-      twilighteboard_oric_write(oric->twilighte, 0x320, 0xff, mkmask);
+      //twilighteboard_oric_write(oric->twilighte, 0x320, 0xff, mkmask);
     }
   }
   else
