+++ plugins/twilighte_board/twilighte.cfg
